Divided the app into multiple files
