+++ webapp/app.py
@@ -1,0 +1,7 @@
+import dash
+
+external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
+
+app = dash.Dash(__name__,
+                external_stylesheets=external_stylesheets,
+                suppress_callback_exceptions=True)
